+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - `Adapter.Odata`:
     - Improve `buildExportExcelURL` function to receive exported excel file URL.
 
+### Fixed
+- `syncer` service:
+    - Fix wrong converting of `'true'` strings to boolean value when performing sync up process.
+
 ## [0.9.0-beta.0] - 2017-02-22
 ### Added
 - `Utils`: